Implement preset management (remove later)

+++ src/insound/presets/Preset.h
@@ -7,10 +7,10 @@ namespace Insound
     struct Preset
     {
         Preset(): name(), volumes() { }
-        Preset(const std::string &name, const std::vector<float> &volumes) :
+        Preset(const std::string &name, const std::vector<double> &volumes) :
             name(name), volumes(volumes) { }
 
         std::string name;
-        std::vector<float> volumes;
+        std::vector<double> volumes;
     };
 }
